src/main/java/com/example/techsupportbot/service/CustomUserDetailsService.java
package com.example.techsupportbot.service;

import com.example.techsupportbot.model.UserEntity;
import com.example.techsupportbot.repository.UserRepository;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.security.core.userdetails.UserDetails;
import org.springframework.security.core.userdetails.UserDetailsService;
import org.springframework.security.core.userdetails.UsernameNotFoundException;
import org.springframework.stereotype.Service;

import java.util.Optional;

/**
 * Service that implements UserDetailsService to provide custom user authentication.
 */
@Service
public class CustomUserDetailsService implements UserDetailsService {

    @Autowired
    private UserRepository userRepository;

    /**
     * Default constructor.
     */
    public CustomUserDetailsService() {}

    /**
     * Loads user details based on the username.
     *
     * @param username the username of the user
     * @return UserDetails the user details object for authentication
     * @throws UsernameNotFoundException if the user is not found
     */
    @Override
    public UserDetails loadUserByUsername(String username) throws UsernameNotFoundException {
        Optional<UserEntity> optionalUser = userRepository.findByUsername(username);
        if (!optionalUser.isPresent()) {
            throw new UsernameNotFoundException("User not found");
        }
        return optionalUser.get();
    }
}
